pyproject.toml
[tool.poetry]
name = "projectdjangocareernavigator"
version = "0.1.0"
description = ""
authors = ["jahoroshi4y <jahoroshi4y@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
